# openapi_pretty.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyDog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/posts/": {
      "post": {
        "summary": "POST /api/posts/",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "tags": [
          "posts"
        ]
      },
      "get": {
        "summary": "GET /api/posts/",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{post_id}": {
      "get": {
        "summary": "특정 게시글의 상세 정보를 조회합니다.",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "summary": "특정 게시글의 내용을 수정합니다. (작성자 본인만 가능)",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "summary": "특정 게시글을 삭제합니다. (작성자 본인만 가능)",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{post_id}/like": {
      "post": {
        "summary": "게시글의 좋아요를 누르거나 취소합니다.",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/users/{author_id}/posts": {
      "get": {
        "summary": "특정 사용자가 작성한 게시물 피드를 페이지네이션으로 조회합니다.",
        "description": "(멍스타그램 전용 프로필 화면의 게시물 목록)",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'author_id'"
          }
        ],
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "components": {
...
    "responses": {
      "UNAUTHORIZED": {
        "description": "인증이 필요하거나 토큰이 유효하지 않습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "UNAUTHORIZED"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "VALIDATION_ERROR": {
        "description": "입력 값이 유효하지 않습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "FORBIDDEN": {
        "description": "권한이 없습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "FORBIDDEN"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "NOT_FOUND": {
        "description": "리소스를 찾을 수 없습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "INTERNAL_ERROR": {
        "description": "서버 내부 오류가 발생했습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
...
    }
  },
...
}